requirements.txt
flask
psycopg2
tensorflow
numpy
Flask-SQLAlchemy===2.1 
Flask-Migrate==1.8.0
pillow
python-swiftclient
python-keystoneclient
